+++ HISTORY.rst
@@ -23,6 +23,7 @@ latest (unreleased)
 
 **Features and Improvements**
 
+* BSQOQ-43: Add controller for disabling an address
 * BSQOQ-34: Add controller for changing the shipping addresses of sale orders
 * BSQOQ-34: Add job to disable change of shipping address on sale orders